[박세환] 리뷰 작성 완료

+++ lib/data/dto/review_request_dto.dart
@@ -1,8 +1,8 @@
 class ReviewWriteDTO {
   final String? content;
-  final double? cleanliness;
-  final double? managementness;
-  final double? friendliness;
+  final int? cleanliness;
+  final int? managementness;
+  final int? friendliness;
 
   ReviewWriteDTO({
     this.content,
